+++ LiquidSimuldation/Math/matrix3x3.h
@@ -7,9 +7,9 @@ class matrix3x3
 public:
 	vector3 i, j, k;
 	matrix3x3();
-	vector2 multiply(const vector2& v, const double& z) const;
-	void scale(const double& scale);
+	vector2 multiply(const vector2& v, const float& z) const;
+	void scale(const float& scale);
 	void scale(const vector3& scale);
-	void rotate(const double& radians);
+	void rotate(const float& radians);
 	void transfer(const vector2& v);
 };
