+++ LiquidSimuldation/Math/vector3.cpp
@@ -10,12 +10,12 @@ vector3::vector3(float x, float y, float z)
 {
 }
 
-double vector3::dot_product(const vector2& v, const double& z) const
+float vector3::dot_product(const vector2& v, const float& z) const
 {
 	return this->x * v.x + this->y * v.y + this->z * z;
 }
 
-void vector3::scale(const double& scale) 
+void vector3::scale(const float& scale) 
 {
 	x *= scale;
 	y *= scale;
